Test FNF, move COG test into STAC test
